refactor: moved agenda to separate component
