+++ .expo/devices.json
@@ -2,7 +2,7 @@
   "devices": [
     {
       "installationId": "360A6B99-5F55-4BA5-A452-5AD9F92D9882",
-      "lastUsed": 1731831212172
+      "lastUsed": 1731832068226
     },
     {
       "installationId": "E799F927-6DD8-4278-B9D2-9559185A79EE",
