+++ components/providers/EventsProvider.tsx
@@ -23,13 +23,12 @@ interface EventsContextType {
     removeEvent: (title: string) => void
 }
 
-//const initialUserEvents: Event[] = getUserEvents("")
 const EventsContext = createContext<EventsContextType | undefined>(undefined)
 
 export function useEvents() {
     const context = useContext(EventsContext)
     if (!context) {
-        throw new Error("useEvents must be used within an EventsProvider");
+        throw new Error("useEvents must be used within an EventsProvider")
     }
     return context
 }
